i18n documentation for required belongs_to message
When you don't explicitly provide `optional: true` to a `belongs_to` definition, rails adds a validation to ensure presence of the `belongs_to` validation.

However, when the validation is added, it is added with a message key of `:required`, which was not documented in the "Translations for Active Record Models" section of the i18n guides.

This commit adds the "required association" validation to the table, so rails users know how to override the message for this validation. The default message is "must exist".